CDA: gforceEvent implemented
Buffer sizes changed to 5 seconds to not throw off CDA with redundant data.
Deceleration is obtained if current speed is within first gear range
and the running average speed past 5 seconds exceeds first gear range.
Makes more use of the entire data rather than comparing T with T-1, also
possible to extend to detect crash earlier (before car reaches speed of
first gear range) by checking deceleration values. This can become useful
during a crash that lasts a long time and takes the car a while to reach
the first gear speed range.

GForce is captured from +-1 the timestamp the critical deceleration is detected.
The maximum g-force value is retrieved in the +-1 range to ensure that it is
the most accurate representative of the vehicle's deceleration. The g-force
is used to confirm the deceleration for a more accurate CDA.