Sensors: Gyroscope removed & Accelerometer + Magnetic field added
Gyroscope causes issues after some period as it begins to drift and output
incorrect values for resulting rotation. Instead, the fixed orientation of the device
is mapped to the vehicle. If at some point the orientation of the device changes
for 5 seconds straight, the fixed orientation of the device is changed.

CDA: Orientation of the vehicle added

If the orientation is alternating constantly, the recent orientation values
are compared with the current fixed position to check if the vehicle has
experienced a critical rotation. If it did, it is confirmed by g-force at that
time period.

Current crash events include collision and overturned.